extract extension method test
